+++ NEWS.md
@@ -1,3 +1,15 @@
+# erahumed 0.14.2
+
+### Modeling advances
+
+* New simulation layer HBD ("Hydrological Balance of Ditches"), with 
+corresponding setup function `setup_hbd()` and `plot.erahumed_hbd()` method. 
+The results from the layer are also included in the Shiny GUI.
+
+### Documentation
+
+* Unified description and title of layers across Shiny app and R documentation. 
+
 # erahumed 0.14.1
 
 ### API changes